[xdl] Remove unused code in src/detach folder
- Remove StandaloneContextService (this code now always runs in user context)
- Remove unused functions in ExponentTools and IosWorkspace
- Remove unused modules
- All of the remaining files in src/detach folder are now TypeScript for easier refactoring
