+++ packages/xdl/src/EmbeddedAssets.ts
@@ -6,7 +6,6 @@ import semver from 'semver';
 import logger from './Logger';
 import * as ExponentTools from './detach/ExponentTools';
 import * as IosPlist from './detach/IosPlist';
-// @ts-ignore IosWorkspace not yet converted to TypeScript
 import * as IosWorkspace from './detach/IosWorkspace';
 import StandaloneContext from './detach/StandaloneContext';
 import { writeArtifactSafelyAsync } from './tools/ArtifactUtils';
