+++ packages/xdl/src/detach/Detach.ts
@@ -3,7 +3,6 @@ import JsonFile from '@expo/json-file';
 import fs from 'fs-extra';
 import { sync as globSync } from 'glob';
 import path from 'path';
-import process from 'process';
 
 import * as EmbeddedAssets from '../EmbeddedAssets';
 import * as UrlUtils from '../UrlUtils';
@@ -12,12 +11,11 @@ import { isDirectory, regexFileAsync } from './ExponentTools';
 import * as IosPlist from './IosPlist';
 import * as IosWorkspace from './IosWorkspace';
 import logger from './Logger';
-import StandaloneBuildFlags from './StandaloneBuildFlags';
 import StandaloneContext from './StandaloneContext';
 
 const SERVICE_CONTEXT_PROJECT_NAME = 'exponent-view-template';
 
-async function ensureBuildConstantsExistsIOSAsync(configFilePath) {
+async function ensureBuildConstantsExistsIOSAsync(configFilePath: string) {
   // EXBuildConstants is included in newer ExpoKit projects.
   // create it if it doesn't exist.
   const doesBuildConstantsExist = fs.existsSync(
@@ -29,13 +27,16 @@ async function ensureBuildConstantsExistsIOSAsync(configFilePath) {
   }
 }
 
-async function _getIosExpoKitVersionThrowErrorAsync(iosProjectDirectory) {
+async function _getIosExpoKitVersionThrowErrorAsync(iosProjectDirectory: string) {
   let expoKitVersion = '';
   const podfileLockPath = path.join(iosProjectDirectory, 'Podfile.lock');
   try {
     const podfileLock = await fs.readFile(podfileLockPath, 'utf8');
     const expoKitVersionRegex = /ExpoKit\/Core\W?\(([0-9.]+)\)/gi;
     const match = expoKitVersionRegex.exec(podfileLock);
+    if (!match) {
+      throw new Error('ExpoKit/Core not found');
+    }
     expoKitVersion = match[1];
   } catch (e) {
     throw new Error(
@@ -45,7 +46,7 @@ async function _getIosExpoKitVersionThrowErrorAsync(iosProjectDirectory) {
   return expoKitVersion;
 }
 
-async function readNullableConfigJsonAsync(projectDir) {
+async function readNullableConfigJsonAsync(projectDir: string) {
   try {
     return getConfig(projectDir);
   } catch (_) {
@@ -53,69 +54,18 @@ async function readNullableConfigJsonAsync(projectDir) {
   }
 }
 
-async function prepareDetachedBuildIosAsync(projectDir, args) {
+async function prepareDetachedBuildIosAsync(projectDir: string, args: any) {
   const config = await readNullableConfigJsonAsync(projectDir);
   if (config && config.exp.name !== SERVICE_CONTEXT_PROJECT_NAME) {
     return prepareDetachedUserContextIosAsync(projectDir, config.exp, args);
   } else {
-    return prepareDetachedServiceContextIosAsync(projectDir, args);
+    throw new Error('Service context is not supported in this version of XDL');
   }
 }
 
-async function prepareDetachedServiceContextIosAsync(projectDir, args) {
-  // service context
-  // TODO: very brittle hack: the paths here are hard coded to match the single workspace
-  // path generated inside IosShellApp. When we support more than one path, this needs to
-  // be smarter.
-  const expoRootDir = path.join(projectDir, '..', '..');
-  const workspaceSourcePath = path.join(projectDir, 'ios');
-  const buildFlags = StandaloneBuildFlags.createIos('Release', { workspaceSourcePath });
-  const context = StandaloneContext.createServiceContext(
-    expoRootDir,
-    null,
-    null,
-    null,
-    /* testEnvironment */ 'none',
-    buildFlags,
-    null,
-    null
-  );
-  const { iosProjectDirectory, supportingDirectory } = IosWorkspace.getPaths(context);
-  const expoKitVersion = await _getIosExpoKitVersionThrowErrorAsync(iosProjectDirectory);
-
-  // use prod api keys if available
-  const prodApiKeys = await _readDefaultApiKeysAsync(
-    path.join(context.data.expoSourcePath, '__internal__', 'keys.json')
-  );
-
-  const { exp } = getConfig(expoRootDir, { skipSDKVersionRequirement: true });
-
-  await IosPlist.modifyAsync(supportingDirectory, 'EXBuildConstants', constantsConfig => {
-    // verify that we are actually in a service context and not a misconfigured project
-    const contextType = constantsConfig.STANDALONE_CONTEXT_TYPE;
-    if (contextType !== 'service') {
-      throw new Error(
-        'Unable to configure a project which has no app.json and also no STANDALONE_CONTEXT_TYPE.'
-      );
-    }
-    constantsConfig.EXPO_RUNTIME_VERSION = expoKitVersion;
-    constantsConfig.API_SERVER_ENDPOINT =
-      process.env.ENVIRONMENT === 'staging'
-        ? 'https://staging.exp.host/--/api/v2/'
-        : 'https://exp.host/--/api/v2/';
-    if (prodApiKeys) {
-      constantsConfig.DEFAULT_API_KEYS = prodApiKeys;
-    }
-    if (exp && exp.sdkVersion) {
-      constantsConfig.TEMPORARY_SDK_VERSION = exp.sdkVersion;
-    }
-    return constantsConfig;
-  });
-}
-
-async function _readDefaultApiKeysAsync(jsonFilePath) {
+async function _readDefaultApiKeysAsync(jsonFilePath: string) {
   if (fs.existsSync(jsonFilePath)) {
-    const keys = {};
+    const keys: any = {};
     const allKeys = await new JsonFile(jsonFilePath).readAsync();
     const validKeys = ['AMPLITUDE_KEY', 'GOOGLE_MAPS_IOS_API_KEY'];
     for (const key in allKeys) {
@@ -128,7 +78,7 @@ async function _readDefaultApiKeysAsync(jsonFilePath) {
   return null;
 }
 
-async function prepareDetachedUserContextIosAsync(projectDir, exp, args) {
+async function prepareDetachedUserContextIosAsync(projectDir: string, exp: any, args: any) {
   const context = StandaloneContext.createUserContext(projectDir, exp);
   const { iosProjectDirectory, supportingDirectory } = IosWorkspace.getPaths(context);
 
@@ -180,7 +130,7 @@ async function prepareDetachedUserContextIosAsync(projectDir, exp, args) {
   }
 }
 
-export async function prepareDetachedBuildAsync(projectDir, args) {
+export async function prepareDetachedBuildAsync(projectDir: string, args: any) {
   if (args.platform === 'ios') {
     await prepareDetachedBuildIosAsync(projectDir, args);
   } else {
@@ -206,7 +156,7 @@ export async function prepareDetachedBuildAsync(projectDir, args) {
 // (see `exponent-view-template.xcodeproj/project.pbxproj` for an example)
 // and `$buildDir/intermediates/assets/$targetPath` on Android (see
 // `android/app/expo.gradle` for an example).
-export async function bundleAssetsAsync(projectDir, args) {
+export async function bundleAssetsAsync(projectDir: string, args: any) {
   const options = await readNullableConfigJsonAsync(projectDir);
   if (!options || options.exp.name === SERVICE_CONTEXT_PROJECT_NAME) {
     // Don't run assets bundling for the service context.
@@ -216,7 +166,7 @@ export async function bundleAssetsAsync(projectDir, args) {
   const bundledManifestPath = EmbeddedAssets.getEmbeddedManifestPath(
     args.platform,
     projectDir,
-    exp
+    exp as any
   );
   if (!bundledManifestPath) {
     logger.warn(
@@ -246,7 +196,7 @@ export async function bundleAssetsAsync(projectDir, args) {
  * that we fetch the resources from the appropriate place when doing builds
  * against self-hosted apps.
  */
-function getExportUrl(manifest) {
+function getExportUrl(manifest: any) {
   const { bundleUrl } = manifest;
   if (bundleUrl.includes(AssetBundle.DEFAULT_CDN_HOST)) {
     return null;
